+++ volmdlr/shapes.py
@@ -48,13 +48,15 @@ from OCP.BRepOffsetAPI import BRepOffsetAPI_MakePipeShell, BRepOffsetAPI_ThruSec
 from OCP.Geom import Geom_Plane
 from OCP.gp import gp_Pnt, gp_Vec, gp_Ax2, gp_Ax1
 from OCP.TopLoc import TopLoc_Location
+from OCP.BRepExtrema import BRepExtrema_DistShapeShape
 import OCP.TopAbs as top_abs  # Topology type enum
 from OCP.TopAbs import TopAbs_Orientation
+
 
 import volmdlr.core_compiled
 from volmdlr import display, edges, surfaces, wires, faces as vm_faces
 from volmdlr.core import edge_in_list
-from volmdlr import to_ocp
+from volmdlr import to_ocp, from_ocp
 from volmdlr.utils.ocp_helpers import plot_edge
 
 
@@ -499,6 +501,24 @@ class Shape(volmdlr.core.Primitive3D):
                              " Try using a closed shell.")
         BRepGProp.VolumeProperties_s(self.wrapped, prop, tol)
         return abs(prop.Mass())
+
+    def distance(self, other: "Shape") -> float:
+        """
+        Minimal distance between two shapes.
+
+        :param other: other shape to calculate distance with.
+        """
+        return BRepExtrema_DistShapeShape(self.wrapped, other.wrapped).Value()
+
+    def distance_points(self, other: "Shape"):
+        """
+        Minimal distance points between two shapes.
+
+        :param other: other shape to calculate distance with.
+        """
+        brep_extrema = BRepExtrema_DistShapeShape(self.wrapped, other.wrapped)
+        return (from_ocp.point3d_from_ocp(brep_extrema.PointOnShape1(N=1)),
+                from_ocp.point3d_from_ocp(brep_extrema.PointOnShape2(N=1)))
 
     def _bool_op(
             self,